chore(gwasUnifiedFlowRefactorSharedFolder): Created seperate files and directories for JobInputModal and SelectConfiguration
