+++ src/Analysis/GWASV2/Steps/ConfigureGWAS/ConfigureGWAS.jsx
@@ -1,29 +1,14 @@
 import React, { useState, useEffect } from 'react';
 import { useMutation } from 'react-query';
 import PropTypes from 'prop-types';
-import { InputNumber, Modal, Input } from 'antd';
-import SelectHareDropDown from '../../Components/SelectHare/SelectHareDropDown';
 import ACTIONS from '../../Utils/StateManagement/Actions';
 import DismissibleMessage from '../../Components/DismissibleMessage/DismissibleMessage';
 import { jobSubmission } from '../../Utils/gwasWorkflowApi';
 import { useSourceContext } from '../../Utils/Source';
 import Congratulations from '../../Components/Congratulations/Congratulations';
+import JobInputModal from '../../Components/JobInputModal/JobInputModal';
+import SelectConfiguration from '../../Components/SelectConfiguration/SelectConfiguration';
 import '../../GWASV2.css';
-import './ConfigureGWAS.css';
-
-const twSudo = {
-  flexCol: {
-    display: 'flex',
-    flexDirection: 'column',
-    margin: 'auto',
-  },
-  flexRow: {
-    display: 'flex',
-    flexDirection: 'row',
-    justifyContent: 'space-between',
-  },
-};
-// todo react tour
 
 const ConfigureGWAS = ({
   dispatch,
@@ -39,8 +24,6 @@ const ConfigureGWAS = ({
 }) => {
   const { source } = useSourceContext();
   const sourceId = source; // TODO - change name of source to sourceId for clarity
-
-  const { flexCol, flexRow } = twSudo;
 
   const [open, setOpen] = useState(false);
   const [showSuccess, setShowSuccess] = useState(false);
@@ -62,17 +45,18 @@ const ConfigureGWAS = ({
   }, [showModal]);
 
   const submitJob = useMutation(
-    () => jobSubmission(
-      sourceId,
-      numOfPCs,
-      covariates,
-      outcome,
-      selectedHare,
-      mafThreshold,
-      imputationScore,
-      selectedCohort,
-      jobName,
-    ),
+    () =>
+      jobSubmission(
+        sourceId,
+        numOfPCs,
+        covariates,
+        outcome,
+        selectedHare,
+        mafThreshold,
+        imputationScore,
+        selectedCohort,
+        jobName
+      ),
     {
       onSuccess: (data) => {
         if (data?.status === 200) {
@@ -83,13 +67,13 @@ const ConfigureGWAS = ({
         } else {
           data.text().then((error) => {
             setErrorText(
-              `GWAS job failed with error: ${JSON.stringify(error)}`,
+              `GWAS job failed with error: ${JSON.stringify(error)}`
             );
             setShowError(true);
           });
         }
       },
-    },
+    }
   );
 
   const handleSubmit = () => {
@@ -115,133 +99,32 @@ const ConfigureGWAS = ({
         />
       )}
       <div className='configure-gwas_container'>
-        <div className='GWASUI-row'>
-          <div className='GWASUI-column'>
-            <label htmlFor='input-numOfPC'>Number of PCs to use</label>
-            <InputNumber
-              id='input-numOfPC'
-              value={numOfPCs}
-              min={1}
-              max={10}
-              onChange={(e) => dispatch({ type: ACTIONS.UPDATE_NUM_PCS, payload: Number(e) })}
-            />
-          </div>
-          <div className='GWASUI-column'>
-            <label htmlFor='input-maf'>MAF Cutoff &nbsp;</label>
-            <InputNumber
-              id='input-maf'
-              value={mafThreshold}
-              onChange={(e) => dispatch({
-                type: ACTIONS.UPDATE_MAF_THRESHOLD,
-                payload: Number(e),
-              })}
-              stringMode
-              step='0.01'
-              min={'0'}
-              max={'0.5'}
-            />
-          </div>
-        </div>
+        <SelectConfiguration
+          numOfPCs={numOfPCs}
+          mafThreshold={mafThreshold}
+          selectedCohort={selectedCohort}
+          covariates={covariates}
+          outcome={outcome}
+          dispatch={dispatch}
+          imputationScore={imputationScore}
+        />
 
-        <div className='GWASUI-row'>
-          <div className='GWASUI-column'>
-            <label htmlFor='input-selectHareDropDown'>HARE Ancestry</label>
-            <SelectHareDropDown
-              id='input-selectHareDropDown'
-              selectedCohort={selectedCohort}
-              covariates={covariates}
-              outcome={outcome}
-              dispatch={dispatch}
-            />
-          </div>
-
-          <div className='GWASUI-column'>
-            <label htmlFor='input-imputation'>
-              Imputation Score Cutoff &nbsp;{' '}
-            </label>
-            <InputNumber
-              id='input-imputation'
-              value={imputationScore}
-              onChange={(e) => dispatch({
-                type: ACTIONS.UPDATE_IMPUTATION_SCORE,
-                payload: Number(e),
-              })}
-              stringMode
-              step='0.1'
-              min={'0'}
-              max={'1'}
-            />
-          </div>
-        </div>
-
-        <Modal
-          okText='Submit'
-          cancelText='Back'
+        <JobInputModal
           open={open}
-          okButtonProps={{ disabled: jobName === '' }}
-          onOk={() => handleSubmit()}
-          onCancel={() => {
-            setOpen(false);
-            dispatch({
-              type: ACTIONS.SET_CURRENT_STEP,
-              payload: 3,
-            });
-          }}
-          title={(
-            <div style={{ ...flexRow, ...{ justifyContent: 'space-between' } }}>
-              <div>Review Details</div>
-            </div>
-          )}
-        >
-          <Input
-            className='gwas-job-name'
-            placeholder='Enter Job Name'
-            onChange={handleEnterJobName}
-          />
-          <div style={flexCol}>
-            <div style={flexRow}>
-              <div>Number of PCs</div>
-              <div>{numOfPCs}</div>
-            </div>
-            <div style={flexRow}>
-              <div>MAF Cutoff</div>
-              <div>{mafThreshold}</div>
-            </div>
-            <div style={flexRow}>
-              <div>HARE Ancestry</div>
-              <div>{selectedHare?.concept_value_name}</div>
-            </div>
-            <div style={flexRow}>
-              <div>Imputation Score Cutoff</div>
-              <div>{imputationScore}</div>
-            </div>
-            <hr />
-            <div style={flexRow}>
-              <div>Cohort</div>
-              <div>{selectedCohort?.cohort_name}</div>
-            </div>
-            <div style={flexRow}>
-              <div>Outcome Phenotype</div>
-              <div>{outcome?.concept_name ?? outcome?.provided_name}</div>
-            </div>
-            {finalPopulationSizes.map((item, key) => (
-              <div key={key} style={flexRow}>
-                <div>{item.population} Size</div>
-                <div>{item.size}</div>
-              </div>
-            ))}
-            <div style={flexRow}>
-              <div>Covariates</div>
-              <div>
-                {covariates.map((covariate, key) => (
-                  <div key={key}>
-                    {covariate?.concept_name ?? covariate.provided_name}
-                  </div>
-                ))}
-              </div>
-            </div>
-          </div>
-        </Modal>
+          jobName={jobName}
+          handleSubmit={handleSubmit}
+          setOpen={setOpen}
+          dispatch={dispatch}
+          handleEnterJobName={handleEnterJobName}
+          numOfPCs={numOfPCs}
+          mafThreshold={mafThreshold}
+          selectedHare={selectedHare}
+          imputationScore={imputationScore}
+          selectedCohort={selectedCohort}
+          outcome={outcome}
+          finalPopulationSizes={finalPopulationSizes}
+          covariates={covariates}
+        />
       </div>
     </div>
   );
